Update 2-viz.html
Made changes to correct following: 
I noticed when looking through Chapter 2 of the course textbook.

The first is the text below Figure 2.10– “...observe that temperatures vary from around 74 grams up to 373 grams.” I imagine that “temperatures” was left from the original Modern Dive text.

Another section that might be tweaked is section 2.7- Boxplots. The concept of outliers is explained clearly in the text, however Figure 2.15 doesn’t have a visual representation of what the outliers look like on a box plot (I suspect there weren’t outliers in the data used to generate the figure). It might be helpful for students to have a visual representation of what outliers look like in addition to a description.